requirements-dev.txt
aspy.yaml==0.3.0
certifi==2017.7.27.1
chardet==3.0.4
coverage==4.4.1
coveralls==1.2.0
decorator==4.1.2
docopt==0.6.2
identify==1.0.5
idna==2.6
ipdb==0.10.3
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
mock==2.0.0
nodeenv==1.2.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
pluggy==0.5.1
pre-commit==0.17.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.4.34
Pygments==2.2.0
pytest==3.2.1
pytest-django==3.1.2
PyYAML==3.12
requests==2.18.4
requirements-tools==1.1.2
setuptools==36.3.0
simplegeneric==0.8.1
tox==2.7.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
